Create a decent working module system
With this setup, I was able to replicate my NixOS config at
https://github.com/foo-dogsquared/nixos-config.
It is initially based from this guy's config at
https://github.com/pmeiyu/guix-config/.
It features neat syntax and (slightly) intuitive workflow, I'd say not
too bad for a first-timer in using GNU Guile.

Speaking of which, in the process of creating this setup, I've learned a
whole lot with Scheme-based languages: hygienic macros, alists, hash
tables, and ways you can create interfaces with those simple things.
It's amazing at how simple yet great at giving you a lot of things to
work on.